+++ lib/widgets/tools/crypto_and_encodings/adfgvx.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:gc_wizard/i18n/app_localizations.dart';
 import 'package:gc_wizard/logic/tools/crypto_and_encodings/adfgvx.dart';
 import 'package:gc_wizard/logic/tools/crypto_and_encodings/polybios.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_alphabetdropdown.dart';
+import 'package:gc_wizard/widgets/common/gcw_alphabetdropdown.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_text_divider.dart';
