+++ lib/widgets/tools/crypto_and_encodings/chao.dart
@@ -3,8 +3,7 @@ import 'package:gc_wizard/i18n/app_localizations.dart';
 import 'package:gc_wizard/logic/tools/crypto_and_encodings/chao.dart';
 import 'package:gc_wizard/logic/tools/crypto_and_encodings/reverse.dart';
 import 'package:gc_wizard/utils/alphabets.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_dropdownbutton.dart';
-import 'package:gc_wizard/widgets/common/base/gcw_alphabetdropdown.dart';
+import 'package:gc_wizard/widgets/common/gcw_alphabetdropdown.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_text_divider.dart';
